Move condensation config helper near condenser (oh-tab-im0z)
