+++ packages/agent-sdk-ts/src/sdk/context/condenser/index.ts
@@ -1,1 +1,2 @@
 export * from './llmSummarizingCondenser';
+export * from './condensationLlmConfig';
